Adding group_path as kwarg to emulate save/load from group

+++ tidy3d/components/data/monitor_data.py
@@ -736,7 +736,7 @@ class DiffractionData(MonitorData, DiffractionDataset):
         fields_norm = {}
         for field_name, field_data in self.field_components.items():
             src_amps = source_spectrum_fn(field_data.f)
-            fields_norm[field_name] = field_data / src_amps
+            fields_norm[field_name] = (field_data / src_amps).astype(field_data.dtype)
 
         return self.copy(update=fields_norm)
